Backend/requirements.txt
flask
flask_cors
supabase
python-dotenv
pytest